Sort selectors by decreasing degree before applying the combining optimization.
Rationale: The number of selectors in a combination is limited by the maximum degree of
any polynomial constraint for those selectors. The rest of the algorithm will attempt to
combine selectors by iterating over them in order. If this order is such that we prefer
to combine selectors for which the maximum degree is about the same, intuitively we
should be able to combine more of them. In practice this leads to slightly better results,
i.e. fewer combinations are typically needed.

Signed-off-by: Daira Hopwood <daira@jacaranda.org>