Change to `Into<Column<Any>>` for pub API methods
As suggested by daira in #345, it's sometimes annoting to call `.into()`
for all of the columns in methods like `enable_equality`.
Specially when you end up with situations where you have `&Column<_>` where you are forced to do something like:
`meta.enable_equality((*column).into())`

This change changes the public API methods that require `Column<Any>`
by `<C: Into<Column<Any>> + Copy>` which removes the need to call `.into()` on the public API and it's called inside of the functions that accept this
parameters.

The `Copy` is added since `Into` consumes the Column, and sometimes we
fall into the situation of needing to clone when we can simply require
the copy and avoid more code verbosity.

Resolves: #345