Remove unnecessary Copy bound & use shadowing
As suggested by @str4d, the `Copy` bound was not necessary in most of
the functions as it should be up to the caller to clone or not.

Also, the diff was larger than it should due to not use shadowing in the
functions that could convert into `Column<Any>` just once.

Co-authored-by: Jack Grigg <jack@electriccoin.co>