fix: StackOverflow caused by Debug recursive call
For the `VerifyFailure::ConstraintNotSatisfied` case we want to customly
impl the `Debug` call so that we can print the annotations within the
`VirtualCell`s.

If we hit any other case of `VerifyFailure` we basically send it to
print and be handled by (ideally Display).

The issue is that the call was forwarded to `Debug` impl again which
ended up in an infinite recursive call to the `Debug` method.
See: https://github.com/privacy-scaling-explorations/halo2/pull/109#issuecomment-1375483235

This is solved by calling `Display` if we hit a case that is not
`ConstraintNotSatisfied`.