simple-example: Failing to verify real proof
The constant value is passed in through the MyCircuit struct.
In MyCircuit::without_witnesses(), this constant is set to F::ZERO
(F::default()). Because keygen uses without_witnesses() to define
the circuit, the constant value is set to zero in ConstraintSystem;
however, at proving time, MyCircuit passes in a nonzero constant.
This causes the equality constraint in assign_advice_from_constant
to fail.

In general, fixed values should be copied in without_witnesses().