Use Vec<F> instead of Vec<Assigned<F>> in prover::commit_phase
Simplify the prover API by receiving the witness as Vec<F> instead of
Vec<Assigned<F>>.  This moves the batch inverstion from the backend to
the frontend, so that the backend directly receives the final witness
values.  This change requires updating the batch_invert_assigned method
to work on Vec instead of Polynomial.